perf: 初期化ロジックの責務分離
